requirements.txt
fastapi
gunicorn
importlib-metadata
numpy
onnx
onnxruntime
Pillow
pydantic
typing
uvicorn

